feat(Winner): 승자에 대한 정보 및 메서드를 domain으로 생성하여 처리함
